PYTHON-5945 Move TestValidateTracingOrNone into test_otel.py
Colocate the temporary tracing-option validator tests with the rest of
the otel-specific, PYTHON-5947-superseded test surface instead of
test_common.py's general validator coverage. Note this moves them
behind the "otel" pytest marker, so they now run alongside the other
otel tests rather than under the default suite.
